implemented sdfatlib20131225 into repo updated documentation about Leonardo's memory limitations vs UNO

+++ SdFat/MinimumSerial.cpp
@@ -26,7 +26,7 @@
  * Do not call this function if you use another serial library.
  * \param[in] baud rate
  */
-void MinimumSerial::begin(unsigned long baud) {
+void MinimumSerial::begin(uint32_t baud) {
   uint16_t baud_setting;
   // don't worry, the compiler will squeeze out F_CPU != 16000000UL
   if (F_CPU != 16000000UL || baud != 57600) {
@@ -44,7 +44,7 @@ void MinimumSerial::begin(unsigned long baud) {
   UBRR0H = baud_setting >> 8;
   UBRR0L = baud_setting;
   // enable transmit and receive
-  UCSR0B |= (1 << TXEN0) | (1 << RXEN0) ;
+  UCSR0B |= (1 << TXEN0) | (1 << RXEN0);
 }
 //------------------------------------------------------------------------------
 /**
